requirements.txt
ultralytics>=8.0.0
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless>=4.8.0
albumentations>=1.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
pytest>=7.0.0
jupyter
ipykernel
